[NB] Use shape coord space for PyMunk bodies
Use shape coordinate space for PyMunk bodies.

This significantly reduces the time required to reset the shape of a
canvas, since the PyMunk bodies do not need to be reconstructed.
Instead, the canvas coordinates are dynamically translated to the shape
coordinate space before conducting a point-query in the `find_shape`
method.